Move other file tree tests into file-tree.spec.ts
